Optimized the job system and made the system generic
The job system has been generailized so the could by applied to npc, enemies and player. refactor the system to only call any event only once and when needed and possible. huge changes on the combat system (currrently not done) and contain basic equipment system that is scalable through unity's game foundation.